edit_rag/requirements.txt
langchain
langchain-community
langchain-nvidia-ai-endpoints
chromadb
python-dotenv
tqdm
redis>=4.0.0
